Milestone 4

+++ Glimpse/app/Http/Controllers/JobController.php
@@ -18,70 +18,82 @@ use App\Models\EducationModel;
 use App\Models\SkillModel;
 use App\Models\JobModel;
 
-class JobController extends Controller
+class JobController
 {
 
     //This method registers a user.
     public function onAddJob(Request $request){
-        
-        $newJob = new JobModel(NULL, request()->get('title'), request()->get('company'), request()->get('description'), request()->get('requirements'));
-        
-        $securityservice = new SecurityService();
-        
-        $didSubmitJob = $securityservice->addJobPosting($newJob);
-        
-        $jobResults = $securityservice->getAllJobPostings();
-        
-        return view('myJobPostings')->with('jobs', $jobResults);
-        
-        if ($didSubmitJob) {
+        if (!empty(Session::get('currentUser'))) {
+            $newJob = new JobModel(NULL, request()->get('title'), request()->get('company'), request()->get('description'), request()->get('requirements'));
             
+            $securityservice = new SecurityService();
+            
+            $didSubmitJob = $securityservice->addJobPosting($newJob);
+            
+            if ($didSubmitJob) {
+                
+            }
+            else {
+                
+            }
+            
+            $jobResults = $securityservice->getAllJobPostings();
+            
+            return view('myJobPostings')->with('jobs', $jobResults);
         }
         else {
-            
+            return view('login');
         }
-        
     }
     
     
     //This method updates the user profile.
     public function updateJob(Request $request){
         
-        $securityservice = new SecurityService();
-        
-        $updateMe = new JobModel(request()->get('editJobID'), request()->get('editTitle'), request()->get('editCompany'), request()->get('editDescription'), request()->get('editRequirements'));
-        
-        $didSubmitJob = $securityservice->UpdateJobPosting($updateMe);
-        
-        $jobResults = $securityservice->getAllJobPostings();
-        
-        return view('myJobPostings')->with('jobs', $jobResults);
-        
-        if ($didSubmitJob) {
+        if (!empty(Session::get('currentUser'))) {
+            $securityservice = new SecurityService();
             
+            $updateMe = new JobModel(request()->get('editJobID'), request()->get('editTitle'), request()->get('editCompany'), request()->get('editDescription'), request()->get('editRequirements'));
+            
+            $didSubmitJob = $securityservice->UpdateJobPosting($updateMe);
+            
+            if ($didSubmitJob) {
+                
+            }
+            else {
+                
+            }
+            
+            $jobResults = $securityservice->getAllJobPostings();
+            
+            return view('myJobPostings')->with('jobs', $jobResults);
         }
         else {
-            
+            return view('login');
         }
     }
     
     
     //This method deletes a user from the database.
     public function deleteJob($jobID){
-        $securityservice = new SecurityService();
-        
-        $didDelete = $securityservice->deleteJobPosting($jobID);
-        
-        $jobResults = $securityservice->getAllJobPostings();
-        
-        return view('myJobPostings')->with('jobs', $jobResults);
-        
-        if ($didDelete) {
+        if (!empty(Session::get('currentUser'))) {
+            $securityservice = new SecurityService();
             
+            $didDelete = $securityservice->deleteJobPosting($jobID);
+            
+            if ($didDelete) {
+                
+            }
+            else {
+                echo didDelete;
+            }
+            
+            $jobResults = $securityservice->getAllJobPostings();
+            
+            return view('myJobPostings')->with('jobs', $jobResults);
         }
         else {
-            echo didDelete;
+            return view('login');
         }
-        
     }
 }
