+++ Glimpse/app/Http/Controllers/PortfolioController.php
@@ -17,27 +17,32 @@ use App\Services\Data\JobHistoryDAO;
 use App\Models\EducationModel;
 use App\Models\SkillModel;
 
-class PortfolioController extends Controller
+class PortfolioController
 {
     //This method registers a user.
     public function onAddJob(Request $request){
-        
-        $newJob = new JobHistoryModel(NULL, request()->get('jobTitle'), request()->get('jobCompany'), request()->get('jobDescription'));
-        
-        $securityservice = new SecurityService();
-        
-        $didSubmitJob = $securityservice->addJobHistory($newJob);
-        
-        $jobResults = $securityservice->getAllJobs();
-        $skillResults = $securityservice->getAllSkills();
-        $educationResults = $securityservice->getAllEducations();
-        return view('portfolio')->with('jobs', $jobResults)->with('skills', $skillResults)->with('educations', $educationResults);
-        
-        if ($didSubmitJob) {
+              
+        if (!empty(Session::get('currentUser'))) {
+            $newJob = new JobHistoryModel(NULL, request()->get('jobTitle'), request()->get('jobCompany'), request()->get('jobDescription'));
             
+            $securityservice = new SecurityService();
+            
+            $didSubmitJob = $securityservice->addJobHistory($newJob);
+            
+            if ($didSubmitJob) {
+                
+            }
+            else {
+                
+            }
+            
+            $jobResults = $securityservice->getAllJobs();
+            $skillResults = $securityservice->getAllSkills();
+            $educationResults = $securityservice->getAllEducations();
+            return view('portfolio')->with('jobs', $jobResults)->with('skills', $skillResults)->with('educations', $educationResults);
         }
         else {
-            
+            return view('login');
         }
         
     }
@@ -46,22 +51,27 @@ class PortfolioController extends Controller
     //This method updates the user profile.
     public function updateJob(Request $request){
         
-        $securityservice = new SecurityService();
-        
-        $updateMe = new JobHistoryModel(request()->get('editJobHistoryID'), request()->get('editJobTitle'), request()->get('editJobCompany'), request()->get('editJobDescription'));
-        
-        $didSubmitJob = $securityservice->updateJob($updateMe);
-        
-        $jobResults = $securityservice->getAllJobs();
-        $skillResults = $securityservice->getAllSkills();
-        $educationResults = $securityservice->getAllEducations();
-        return view('portfolio')->with('jobs', $jobResults)->with('skills', $skillResults)->with('educations', $educationResults);
-        
-        if ($didSubmitJob) {
+        if (!empty(Session::get('currentUser'))) {
+            $securityservice = new SecurityService();
             
+            $updateMe = new JobHistoryModel(request()->get('editJobHistoryID'), request()->get('editJobTitle'), request()->get('editJobCompany'), request()->get('editJobDescription'));
+            
+            $didSubmitJob = $securityservice->updateJob($updateMe);
+            
+            if ($didSubmitJob) {
+                
+            }
+            else {
+                
+            }
+            
+            $jobResults = $securityservice->getAllJobs();
+            $skillResults = $securityservice->getAllSkills();
+            $educationResults = $securityservice->getAllEducations();
+            return view('portfolio')->with('jobs', $jobResults)->with('skills', $skillResults)->with('educations', $educationResults);
         }
         else {
-
+            return view('login');
         }
     }
     
@@ -70,20 +80,26 @@ class PortfolioController extends Controller
 
     //This method deletes a user from the database.
     public function deleteJob($jobHistoryID){
-        $securityservice = new SecurityService();
         
-        $didDelete = $securityservice->deleteJob($jobHistoryID);
-        
-        $jobResults = $securityservice->getAllJobs();
-        $skillResults = $securityservice->getAllSkills();
-        $educationResults = $securityservice->getAllEducations();
-        return view('portfolio')->with('jobs', $jobResults)->with('skills', $skillResults)->with('educations', $educationResults);
-        
-        if ($didDelete) {
+        if (!empty(Session::get('currentUser'))) {
+            $securityservice = new SecurityService();
             
+            $didDelete = $securityservice->deleteJob($jobHistoryID);
+            
+            if ($didDelete) {
+                
+            }
+            else {
+                echo didDelete;
+            }
+            
+            $jobResults = $securityservice->getAllJobs();
+            $skillResults = $securityservice->getAllSkills();
+            $educationResults = $securityservice->getAllEducations();
+            return view('portfolio')->with('jobs', $jobResults)->with('skills', $skillResults)->with('educations', $educationResults);
         }
         else {
-            echo didDelete;
+            return view('login');
         }
         
     }
@@ -92,23 +108,28 @@ class PortfolioController extends Controller
     //This method registers a user.
     public function onAddEducation(Request $request){
         
-        $newEducation = new EducationModel(NULL, request()->get('educationSchoolName'), request()->get('educationDegree'), request()->get('educationFieldOfStudy'), request()->get('educationStartDate'), request()->get('educationEndDate'));
-        
-        $securityservice = new SecurityService();
-        
-        $didSubmitJob = $securityservice->addEducation($newEducation);
-        
-        $securityservice = new SecurityService();
-        $jobResults = $securityservice->getAllJobs();
-        $skillResults = $securityservice->getAllSkills();
-        $educationResults = $securityservice->getAllEducations();
-        return view('portfolio')->with('jobs', $jobResults)->with('skills', $skillResults)->with('educations', $educationResults);
-        
-        if ($didSubmitJob) {
+        if (!empty(Session::get('currentUser'))) {
+            $newEducation = new EducationModel(NULL, request()->get('educationSchoolName'), request()->get('educationDegree'), request()->get('educationFieldOfStudy'), request()->get('educationStartDate'), request()->get('educationEndDate'));
             
+            $securityservice = new SecurityService();
+            
+            $didSubmitJob = $securityservice->addEducation($newEducation);
+            
+            if ($didSubmitJob) {
+                
+            }
+            else {
+                
+            }
+            
+            $securityservice = new SecurityService();
+            $jobResults = $securityservice->getAllJobs();
+            $skillResults = $securityservice->getAllSkills();
+            $educationResults = $securityservice->getAllEducations();
+            return view('portfolio')->with('jobs', $jobResults)->with('skills', $skillResults)->with('educations', $educationResults);
         }
         else {
-            
+            return view('login');
         }
         
     }
@@ -117,22 +138,27 @@ class PortfolioController extends Controller
     //This method updates the user profile.
     public function updateEducation(Request $request){
         
-        $securityservice = new SecurityService();
-        
-        $updateMe = new EducationModel(request()->get('editEducationID'), request()->get('editEducationSchoolName'), request()->get('editEducationDegree'), request()->get('editEducationFieldOfStudy'), request()->get('editEducationStartDate'), request()->get('editEducationEndDate'));
-        
-        $didSubmitJob = $securityservice->updateEducation($updateMe);
-        
-        $jobResults = $securityservice->getAllJobs();
-        $skillResults = $securityservice->getAllSkills();
-        $educationResults = $securityservice->getAllEducations();
-        return view('portfolio')->with('jobs', $jobResults)->with('skills', $skillResults)->with('educations', $educationResults);
-        
-        if ($didSubmitJob) {
+        if (!empty(Session::get('currentUser'))) {
+            $securityservice = new SecurityService();
             
+            $updateMe = new EducationModel(request()->get('editEducationID'), request()->get('editEducationSchoolName'), request()->get('editEducationDegree'), request()->get('editEducationFieldOfStudy'), request()->get('editEducationStartDate'), request()->get('editEducationEndDate'));
+            
+            $didSubmitJob = $securityservice->updateEducation($updateMe);
+            
+            if ($didSubmitJob) {
+                
+            }
+            else {
+                
+            }
+            
+            $jobResults = $securityservice->getAllJobs();
+            $skillResults = $securityservice->getAllSkills();
+            $educationResults = $securityservice->getAllEducations();
+            return view('portfolio')->with('jobs', $jobResults)->with('skills', $skillResults)->with('educations', $educationResults);
         }
         else {
-            
+            return view('login');
         }
     }
     
@@ -141,20 +167,26 @@ class PortfolioController extends Controller
     
     //This method deletes a user from the database.
     public function deleteEducation($educationID){
-        $securityservice = new SecurityService();
         
-        $didDelete = $securityservice->deleteEducation($educationID);
-        
-        $jobResults = $securityservice->getAllJobs();
-        $skillResults = $securityservice->getAllSkills();
-        $educationResults = $securityservice->getAllEducations();
-        return view('portfolio')->with('jobs', $jobResults)->with('skills', $skillResults)->with('educations', $educationResults);
-        
-        if ($didDelete) {
+        if (!empty(Session::get('currentUser'))) {
+            $securityservice = new SecurityService();
             
+            $didDelete = $securityservice->deleteEducation($educationID);
+            
+            if ($didDelete) {
+                
+            }
+            else {
+                echo didDelete;
+            }
+            
+            $jobResults = $securityservice->getAllJobs();
+            $skillResults = $securityservice->getAllSkills();
+            $educationResults = $securityservice->getAllEducations();
+            return view('portfolio')->with('jobs', $jobResults)->with('skills', $skillResults)->with('educations', $educationResults);
         }
         else {
-            echo didDelete;
+            return view('login');
         }
         
     }
@@ -162,66 +194,79 @@ class PortfolioController extends Controller
     //This method registers a user.
     public function onAddSkill(Request $request){
         
-        $newSkill = new SkillModel(NULL, request()->get('skill'));
-        
-        $securityservice = new SecurityService();
-        
-        $didSubmitJob = $securityservice->addSkill($newSkill);
-        
-        $jobResults = $securityservice->getAllJobs();
-        $skillResults = $securityservice->getAllSkills();
-        $educationResults = $securityservice->getAllEducations();
-        return view('portfolio')->with('jobs', $jobResults)->with('skills', $skillResults)->with('educations', $educationResults);
-        
-        if ($didSubmitJob) {
+        if (!empty(Session::get('currentUser'))) {
+            $newSkill = new SkillModel(NULL, request()->get('skill'));
             
+            $securityservice = new SecurityService();
+            
+            $didSubmitJob = $securityservice->addSkill($newSkill);
+            
+            if ($didSubmitJob) {
+                
+            }
+            else {
+                
+            }
+            
+            $jobResults = $securityservice->getAllJobs();
+            $skillResults = $securityservice->getAllSkills();
+            $educationResults = $securityservice->getAllEducations();
+            return view('portfolio')->with('jobs', $jobResults)->with('skills', $skillResults)->with('educations', $educationResults);
         }
         else {
-            
-        }
-        
+            return view('login');
+        }       
     }
     
     
     //This method updates the user profile.
     public function updateSkill(Request $request){
-        
-        $securityservice = new SecurityService();
-        
-        $updateMe = new SkillModel(request()->get('editSkillID'), request()->get('editSkill'));
-        
-        $didSubmitJob = $securityservice->updateSkill($updateMe);
-        
-        $jobResults = $securityservice->getAllJobs();
-        $skillResults = $securityservice->getAllSkills();
-        $educationResults = $securityservice->getAllEducations();
-        return view('portfolio')->with('jobs', $jobResults)->with('skills', $skillResults)->with('educations', $educationResults);
-        
-        if ($didSubmitJob) {
+        if (!empty(Session::get('currentUser'))) {
+            $securityservice = new SecurityService();
             
+            $updateMe = new SkillModel(request()->get('editSkillID'), request()->get('editSkill'));
+            
+            $didSubmitJob = $securityservice->updateSkill($updateMe);
+            
+            if ($didSubmitJob) {
+                
+            }
+            else {
+                
+            }
+            
+            $jobResults = $securityservice->getAllJobs();
+            $skillResults = $securityservice->getAllSkills();
+            $educationResults = $securityservice->getAllEducations();
+            return view('portfolio')->with('jobs', $jobResults)->with('skills', $skillResults)->with('educations', $educationResults);
         }
         else {
-            
+            return view('login');
         }
     }
     
     
     //This method deletes a user from the database.
     public function deleteSkill($skillID){
-        $securityservice = new SecurityService();
-        
-        $didDelete = $securityservice->deleteSkill($skillID);
-        
-        $jobResults = $securityservice->getAllJobs();
-        $skillResults = $securityservice->getAllSkills();
-        $educationResults = $securityservice->getAllEducations();
-        return view('portfolio')->with('jobs', $jobResults)->with('skills', $skillResults)->with('educations', $educationResults);
-        
-        if ($didDelete) {
+        if (!empty(Session::get('currentUser'))) {
+            $securityservice = new SecurityService();
             
+            $didDelete = $securityservice->deleteSkill($skillID);
+            
+            if ($didDelete) {
+                
+            }
+            else {
+                echo didDelete;
+            }
+            
+            $jobResults = $securityservice->getAllJobs();
+            $skillResults = $securityservice->getAllSkills();
+            $educationResults = $securityservice->getAllEducations();
+            return view('portfolio')->with('jobs', $jobResults)->with('skills', $skillResults)->with('educations', $educationResults);
         }
         else {
-            echo didDelete;
+            return view('login');
         }
         
     }
